pyproject.toml
[tool.poetry]
name = "website-tools"
version = "0.1.0"
description = ""
authors = ["Nikolaus Gullotta <nikolaus.gullotta@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
Pillow = "^10.0.1"


[tool.poetry.group.dev.dependencies]
mdformat-gfm = "^0.3.5"
isort = "^5.12.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79